Make YubiKeyDeviceListener resettable
Currently, YubiKeyDeviceListener only exposes a singleton instance with no
way for a consumer to reset it once it is started. The constructor is private, so
the consumer has no option to use a private instance.

This adds the ability for the consumer to reset the lazy-initialized singleton so that
the listen thread doesn't stick around.

Also, even though this class implemented IDisposable, most things that need to be
in the Dispose method weren't there, including a way to stop the thread.

The listen thread is now replaced by an asynchronous task that doesn't consume
a system thread while it's not working.